MIX-262 refactor: extract tracklist game business logic into useTracklistGame hook
